lms-back-Money 1/lms-back Original/lms-back/src/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==5.2.6
django-allauth==0.54.0
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
google-auth==2.40.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.2.0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.10.0
zope.interface==8.0
weasyprint==62.3
tinycss2==1.3.0
cssselect2==0.7.0
